More int size tweaks to avoid build warnings with VC++ on x64
git-svn-id: svn://svn.gna.org/svn/mysqlpp/trunk@2560 5946f024-35f4-0310-b1b3-9a3be3380cfb

+++ lib/field_names.cpp
@@ -39,7 +39,7 @@ namespace internal { extern void str_to_lwr(std::string& s); }
 void
 FieldNames::init(const ResultBase* res)
 {
-	int num = res->num_fields();
+	size_t num = res->num_fields();
 	reserve(num);
 
 	for (int i = 0; i < num; i++) {
